refactor: update fetchPlaceholders import path and enhance multi-store support
