+++ scripts/initializers/checkout.js
@@ -1,7 +1,7 @@
 import { initializers } from '@dropins/tools/initializer.js';
 import { initialize, setFetchGraphQlHeaders } from '@dropins/storefront-checkout/api.js';
 import { initializeDropin } from './index.js';
-import { fetchPlaceholders } from '../aem.js';
+import { fetchPlaceholders } from '../scripts.js';
 import { getHeaders } from '../configs.js';
 
 await initializeDropin(async () => {
